Not sure something though

+++ helpers.h
@@ -3,6 +3,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
+#include <time.h>
 
 #define BUFSIZE 1024
 
@@ -10,5 +11,5 @@ struct sockaddr_un *pet__abstract_socket(const char *name, int len, int *addr_le
 
 struct pet {
     char name[32];
-    int age;
+    float age;
 };
